Clarify the data flow between the different threads
